+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- init handler with exception suppression
+
 ### Changed
 
+- Added init handler with exception suppression
 - [playground-quarkus] set docker image to eclipse-temurin:21.0.1_12-jdk
 - [playground-quarkus] validation disabled for FTLX
 - [playground-quarkus] better validation handling